feat: Update project configuration and dependencies
This commit updates the project's package name to `hostrine-admin`, refactors the build process to include TypeScript compilation, and adjusts various dependencies to their latest stable versions. Additionally, it updates the `tsconfig.json` and `vite.config.ts` files for better compatibility and builds, and changes the default language in `index.html` to Portuguese (pt-BR).